requirements.txt
easyocr>=1.7.1
pdf2image>=1.16.3
pytesseract>=0.3.10
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
spacy>=3.7.2
transformers>=4.35.2
torch>=2.1.1
Pillow>=10.1.0
numpy>=1.26.2
python-dotenv>=1.0.0
